Add anomaly detection demo notebook and utils. Update Pipfile
The notebook in this commit is almost same as the ci/cd demo
notebook in the internal repo. The only changes are as follows
- remove query that identifies ci cd deployments internally, for
  privacy reasons
- update cache dir
- update markdown
- add grafana dashboard screenshot for a ci cluster, as the
  dashboard is not accessible outside VPN

utils.py is almost the same as prom_query.py used internally.
Only difference is that results are stored as parquet instead of
pickle format.

evaluate.py is a subset of utils.py used internally.